requirements.txt
torch
transformers
langchain cli
faiss-cpu
numpy
